Домашнее задание 5-1
